fix: content maintenance
